% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOStutorials-package.R
\docType{package}
\name{FOStutorials-package}
\alias{FOStutorials}
\alias{FOStutorials-package}
\title{FOStutorials: Learnr Tutorials for Multivariate Data Analysis in the LUMC FOS Course}
\description{
Interactive learnr tutorials used in the class Multivariate Data Analysis of the FOS course of the LUMC. The tutorials cover principal component analysis (PCA) on the iris and food data sets, and PCA, PCA-class and PLS-DA on an NMR metabolomics data set.
}
\author{
\strong{Maintainer}: Rico Derks \email{r.j.e.derks@lumc.nl}

Authors:
\itemize{
  \item Rico Derks \email{r.j.e.derks@lumc.nl}
  \item Aswin Verhoeven
}

}
\keyword{internal}
