% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tutorial.R
\name{list_fos_tutorials}
\alias{list_fos_tutorials}
\title{List the FOS tutorials}
\usage{
list_fos_tutorials()
}
\value{
A data frame with the tutorial \code{name}, \code{title} and \code{description}.
}
\description{
Lists the learnr tutorials included in this package.
}
\examples{
list_fos_tutorials()
}
