% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tutorial.R
\name{run_fos_tutorial}
\alias{run_fos_tutorial}
\title{Run a FOS tutorial}
\usage{
run_fos_tutorial(name, ...)
}
\arguments{
\item{name}{Name of the tutorial, e.g. \code{"Iris-dataset"}, \code{"Food-dataset"}
or \code{"NMR_metabolomics"}.}

\item{...}{Further arguments passed on to \code{\link[learnr:run_tutorial]{learnr::run_tutorial()}}.}
}
\value{
Called for its side effect of starting the tutorial.
}
\description{
Starts one of the learnr tutorials of the Multivariate Data Analysis class
in your browser (or in the RStudio viewer). Use \code{\link[=list_fos_tutorials]{list_fos_tutorials()}} to
see which tutorials are available.
}
\examples{
\dontrun{
run_fos_tutorial("Iris-dataset")
}
}
